[Perform] 달력이 반영 버그 수정중
- 컴포넌트 m2_board와 m2_tabbar 삭제. 달력을 setState로 봐야할 거 같아서 한 파일로 제작 성공적으로 수정 후에 최적화 예정.
- 현 상태 고르면, 퍼센테이지 반영. 다만, 달력이 제대로 반영되고 있지 않음.
